1、接口

+++ app/src/main/java/com/dog/manage/app/model/Dog.java
@@ -13,7 +13,7 @@ public class Dog implements Serializable {
     private Integer lincenceId = 0;
     private Integer licenceStatus;
     private String dogType;
-    private long dogAge = 0;
+    private Integer dogAge = 0;
     private String createdTime;
     private Integer userId;
     private String userName;
